fix: Grid unSelectRow no longer messes up internal selection state
The "selected" property is an array of the selected row's indices.

Previously the unSelectRow method wrote DOM elements instead of their index to the array.
This caused trouble later when trying to work  with the selected rows.

This fix removes this faulty behaviour.
It removes the index of the unselected row from the array.

(cherry picked from commit 436dcacea185fd48af8be193a7ac5f697b910bc2)