fix(index.php): prevent empty host variable in "mailto" logic
If "[at]" was not contained in the "$_REQUEST['_url']", assigning "$host" would output a warning.
This happened because "explode" returned an array with just one element.
Thus, "$host" could not be assigned.

The new logic makes it explicit, that "[mailto]" and "[at]" have to be contained in "$_REQUEST['_url']".
If that's the case, "$user" and "$host" will always get a value assigned that is at least an empty string.
The "!empty()" check ensures that the header is only set when the variables are not an empty string.