fix: media upload: unzip does not work
This commit adjusts the formatting of the upload/Manager.js code. This included updates for
function declarations, adding spaces between keywords and parentheses. There were changes in the
logic inside the uploadFiles, and injectForm functions. Also, unnecessary lines of code were
removed, and some deprecated functionalities were cleared out. The git commit encapsulates
substantial code format updates to the upload and form management functions in an effort to
increase code readability and maintainability. The changes in this commit should not affect the
functionality of the code.

Related: quiqqer/core#1422