fix(phpstan): ensure object is an instance of `Folder` and modify `storeData` duration time
In `admin/ajax/media/copy.php`, an additional check is included to ensure that the object `$Folder`
is an instance of `Folder`. If it isn't, then throws an exception with a localised message.

Moreover, in `admin/ajax/system/cache/mongoCheck.php`, the duration parameter in
`$CacheDriver->storeData` was changed from 'false' to '10', which means now the data will be stored
for 10 minutes.