fix(phpstan): add missing class existence checks in export.php and Nobody.php
Updated several scripts to include checks for required classes. In export.php, added a check for
the existence of 'QUI\\HtmlToPdf\\Document'. If the class doesn't exist, an error is logged and
execution stops. Likewise, in Nobody.php now checks for the existence of
'QUI\\ERP\\Currency\\Handler'. If the class doesn't exist, it throws an exception. These changes
aim to make the scripts more robust and debug-friendly.