feat: skip file system check flag for the console update command
An option to skip file-system-check during an update has been added to the update console tool.
The decision on whether to execute a filesystem check or not is now dependent on this new
'skip-filesystem-check' argument. `./console update --skip-filesystem-check`

The corresponding localization for the new argument has been added to the de.xml and en.xml files.

Related: quiqqer/core#1440