fix: remove redundant 'deleted' field from User.php
The 'deleted' field was being set to `1` every time a user was disabled. This field was redundant,
as the 'active' field was being set to `-1`, fully representing the disabled state of the user. To
simplify and declutter the code, the 'deleted' field has been removed from `src/QUI/Users/User.php`.

Related: quiqqer/core#1438
Related: quiqqer/core#1442