fix: update 2fa mail sending and otp generation code
The purpose of this commit is to adjust the mail for the 2FA and OTP generation process. The mail
sending functionality has been adjusted to send the OTP code to the current session user instead of
initiating the password reset process. Additionally, the generation of the OTP has also been
modified to use a different method. The new OTP is now randomly generated using a for loop, instead
of the previous method where it was simply padded with zeros.