feat: add cli system backup functionality
This commit introduces a new file, `Backup.php`, under the `QUI\System\Console\Tools` namespace.
The file contains a `Backup` class, which provides a command-line interface for creating backups of
the whole system, the database, or the filesystem.

Backup types can be specified when calling the class with the 'full', 'db', or 'filesystem'
argument. The class automatically handles exceptions and logs helpful error messages while creating
backups. This feature increases the reliability of the system by allowing easy creation of backups
when needed.

The `Backup` class method `createDatabaseBackup` generates database backups with the `mysqldump`
utility, and `createFilesystemBackup` generates filesystem backups with the `tar` utility. The
backups are saved in a dedicated backup folder.