fix: add event firing to upload manager
Modified the upload manager to include event firing functionality. A new 'fileUploadRefresh' event
is now fired when the upload refreshes. Additionally, a 'fileUploadRefresh' event is now also fired
on the completion of each upload part with the progress percent as the parameter. The 'onRefresh'
event handler has been added to the existing event handlers.