fix: use srcset for image HTML generation
This commit simplifies the image HTML generation process in the `Utils` class. We have refactored
the process by which picture source sets are built, renamed `imageX` variable to `imageScale` to
improve readability and made a few optimization and cleanup.

The srcset is now an array for easier manipulation and the mime type of the image is also captured.
The way image set entries are created and handled has been simplified, removing redundant code and
conditions. The commit also includes changes to ensure a single picture instead of multiple
sources.

The changes aim to improve performance and code readability. Modifications are focused around the
`getImageHTML` method.