fix: improve server request handling and vhost parsing
Fixes made to the handling of server requests and the parsing of vhosts. Server request now checks
whether `$_SERVER['HTTP_HOST']` is set before use, defaulting to `null` if not. If set, the
relevant vhost data is directly accessed.

Wildcards (*) are workable (beta and trial version)

Added a check for `*` in vhost keys and a wildcard matching for `httpHost`, which makes the
handling of various vhost configurations more flexible.

A switch from `pathinfo` to `pathInfo` to match usual camelCase naming convention.

Lastly, refactored repetitive checks on `$_SERVER['HTTP_HOST']` and `vhosts` to use a `$vhost`
variable instead, making the code more DRY.