fix(Upload + BulkUpload): enhance error handling + standardize function declaration whitespace
This commit standardizes function declaration in `BulkUpload.js`, `Form.js`, and `Manager.js`
scripts to include a space before the parentheses for anonymous functions, improving overall
readability and consistency.

Additionally, error handling is enhanced significantly. In `BulkUpload.js`, we now search for tags
precisely and trigger an error if tags are not found. Responding to this, an error event has been
introduced.

Moreover, in `Form.js`, an optional flag has been added to decide whether errors are to be shown
using the message handler or not. Also, the `error` event receiving modified arguments, now
includes upload manager instances.

Lastly, in `Manager.js`, the uploader now triggers an error, when encountered during file upload.