fix: improve layout and animations of login control
Improve the layout of the login page by grouping social logins and form logins into separate
sections. Additionally, add subtle animations to improve the experience of users when they're
logging in or registering. This results in a more organized layout and makes the page more visually
appealing.

Other changes include:
- Remove unnecessary author comment lines
- Improve code formatting
- Autofocus on the email field when reset password is clicked.