refactor: improve user interface and backend usability
This commit includes major changes in the admin template and User class. It restructures and adds
additional classes to the 'authenticators' table in `admin/template/users/security.html` while also
introducing improved copy and localization of text regarding 2FA authentication in the locale xml
files.

As for the User class, ES6 fat arrow functions are used to replace traditional function syntax for
better lexical `this` binding. Additionally, the older `var` is replaced with `let` for defining
variables. It also introduces some changes in Promise structure and makes multiple improvements to
function definitions for enhance code readability and consistency.

Users of the admin panel should see a more user-friendly and understandable user interface after
these changes. Changes in the backend, particularly in the User class, provide better asynchronous
handling of events and stylistic improvements in line with modern JavaScript programming practices.