fix: handle console script name validation and directory change in quiqqer.php
The commit includes a fix where it checks if the script name is in the list of valid scripts. If
not, it checks if the basename of the script name is 'console' and changes the directory to CMS_DIR
if it's defined. This ensures the script runs in the correct context.


(cherry picked from commit e3986bcdd715ca41dc46796fad656ad39ebca316)

Co-authored-by: Henning <leutz@pcsg.de>