fix: enhanced security settings UI and updated locale files
Updated the `users/security.html` file to add a new table for additional authentication settings
and an info on available two-factor authentications. The changes also include condition handling
for the absence of two-factor authentication and user-specific notice regarding the same.

In sync with the UI enhancements, existing entries in the German (`de.xml`) and English (`en.xml`)
locale files were updated and new ones were added to accommodate the requirements of the new UI
elements such as additional authentication settings, availability of two-factor authentication, and
user-specific notices.