fix: enhance user session checks and 2FA handling
In this commit, we made several important changes to how users are authenticated and their sessions
are managed.

1. In `src/QUI/Users/Manager.php`, we did a major overhaul of how user sessions are checked. Now,
instead of just checking if the 'uid' and 'auth' session variables exist, we've added a new check
to verify the state of the secondary login type. This involves checking if two-factor
authentication (2FA) is enabled, if it’s possible, or if it’s not at all. We’ve also made
changes to how a session is destroyed, only doing so if the user isn't authenticated.

2. In `src/QUI/Users/Auth/VerifiedMail2FA.php`, instead of just silently failing and returning when
a user ID isn't found in a session, an exception is now thrown.

3. A new function `has2FAAuthenticator()` was implemented in `src/QUI/Users/Utils.php`. This
function checks if a user has at least one secondary authenticator.

4. An additional constant 'QUIQQER_BACKEND' was defined in `admin/index.php`.

These changes aim to enhance the security of user session handling and improve 2FA flow across the
system.