fix: updated iv empty check in encryption
This commit changes the condition for initializing IV in the encryption module. Previously, the IV
was being initialized only if it was explicitly set to false. This could potentially lead to issues
if the IV is somehow empty (but not exactly 'false'). To address this, the condition has been
updated to initialize the IV if it is empty. This would cover all cases where the IV is not valid
(including false, NULL, empty string etc).

Related: quiqqer/core#1488