fix: add localization to system update timestamps
This commit includes additional code to the System.js file to enable localization of system update
timestamps. Specifically, this includes the addition of a new 'language' variable set to 'en-US' by
default, and the introduction of a switch case for other language options (currently 'de'). The
time and date options have been specified with an 'options' object. These modifications affect how
the 'lastUpdateCheck' and 'lastUpdate' variables are formatted and displayed.