fix: update smtp authentication logic
The commit updates the SMTP authentication logic in the `admin/ajax/system/mailTest.php`. It now
checks the 'SMTP' parameter to determine whether to set `SMTPAuth` to true, whereas, earlier, it
was defaulted to true whenever 'SMTPServer', 'SMTPUser', 'SMTPPass' or 'SMTPPort' parameters were
not empty.

This commit also includes an update in 'bin/QUI/controls/system/settings/Mail.js', where the 'SMTP'
checkbox is being accounted for. The code related to clearing parameters if the 'SMTP' checkbox is
not checked, has been commented out. This needs further review as it can have a significant impact
on the application's behavior.

Related: quiqqer/core#1492