fix: improve smtp configuration handling
This commit restructures and improves SMTP configuration handling in the mailTest.php file. Instead
of using a flat structure, the configuration parameters are now encompassed in a single conditional
for SMTP. This improves readability and reduces redundancy in the script. Further, this change
ensures that each SMTP setting is only set if the corresponding parameter is not empty in the
`params` array. This is done to avoid unwanted overwriting of defaults or other settings.

Related: quiqqer/core#1492