fix: update to support uuid in permissions manager
Updated Permissions Manager to support the use of UUIDs for users and groups. Previously, the
identifier used for the user and group was a number and now it can also be a UUID.
`Ramsey\Uuid\Uuid` is used to verify whether the supplied identifier is a valid UUID. The code
checks if the supplied id is a valid UUID and if not, it falls back to the numerical id.

The `checkPermissionList` method in `Permission.php` has also been updated to support UUIDs. It
checks if the identifier starts with 'u' (for user) or 'g' (for group) and then compares the
identifier with the User's UUID.

Related: quiqqer/core#1494